+++ lib/src/http/timeout_config.dart
@@ -1,19 +1,17 @@
 ///Responsible for configuring the default network layer timeout settings.
 class TimeoutConfig {
   TimeoutConfig({
-    required int connectionTimeout,
-    required int receiveTimeout,
-  })  : _connectionTimeout = connectionTimeout,
-        _receiveTimeout = receiveTimeout;
+     int? connectionTimeout,
+     int? receiveTimeout,
+  })  : _connectionTimeout = connectionTimeout ?? 30,
+        _receiveTimeout = receiveTimeout ?? 30;
 
   final int _connectionTimeout;
   final int _receiveTimeout;
 
-  final int _milliseconds = 1000;
+  ///Set the timeout of a request in seconds
+  Duration get connectionTimeout => Duration(seconds: _connectionTimeout);
 
-  ///Set the timeout of a request in milliseconds
-  int get connectionTimeout => _connectionTimeout * _milliseconds;
-
-  ///Set the timeout in milliseconds
-  int get receiveTimeout => _receiveTimeout * _milliseconds;
+  ///Set the timeout in seconds
+  Duration get receiveTimeout => Duration(seconds: _receiveTimeout);
 }