+++ CHANGELOG.md
@@ -73,3 +73,13 @@
 
 - DIO dependency update;
 - Adding personality error return in mock.
+
+- Fix: Now when a request is made a notify is send to all listeners
+
+## [0.0.15] - Add cacheExpiresIn field
+
+- Added adding field cacheExpiresIn
+
+## [0.0.16] - Update lib
+
+- DIO dependency update;